+++ src/relay.js
@@ -43,7 +43,7 @@ function handleConnection(ws) {
       const payload = JSON.stringify(msg);
       for (const client of clients) {
         if (
-          (client.role === 'meta-preview' || client.role === 'banter-inject' || client.role === 'tui-preview' || client.role === 'sdf2d-preview' || client.role === 'sdf3d-preview' || client.role === 'bantervr-sdf3d-inject')
+          (client.role === 'meta-preview' || client.role === 'banter-inject' || client.role === 'altspace-inject' || client.role === 'tui-preview' || client.role === 'sdf2d-preview' || client.role === 'sdf3d-preview' || client.role === 'bantervr-sdf3d-inject')
           && client.readyState === WebSocket.OPEN
         ) {
           client.send(payload);